Re-write top-level interface and inter-proxy communication to account for the change in PoolSpecs struct

+++ contracts/callpaths/HotPath.sol
@@ -32,8 +32,7 @@ contract HotPath is MarketSequencer, SettleLayer, PoolRegistry, ProtocolAccount 
     
     function swapExecute (address base, address quote,
                           uint256 poolIdx, bool isBuy, bool inBaseQty, uint128 qty,
-                          uint24 poolTip,
-                          uint128 limitPrice, uint128 minOutput,
+                          uint16 poolTip, uint128 limitPrice, uint128 minOutput,
                           uint8 reserveFlags) internal
         returns (int128){
         
@@ -69,7 +68,7 @@ contract HotPath is MarketSequencer, SettleLayer, PoolRegistry, ProtocolAccount 
     }
     
     function preparePoolCntx (address base, address quote,
-                              uint256 poolIdx, uint24 poolTip,
+                              uint256 poolIdx, uint16 poolTip,
                               bool isBuy, bool inBaseQty, uint128 qty) private
         returns (PoolSpecs.PoolCursor memory) {
         PoolSpecs.PoolCursor memory pool = queryPool(base, quote, poolIdx);
@@ -82,10 +81,10 @@ contract HotPath is MarketSequencer, SettleLayer, PoolRegistry, ProtocolAccount 
 
     function swapEncoded (bytes calldata input) internal returns (int128 outFlow) {
         (address base, address quote,
-         uint256 poolIdx, bool isBuy, bool inBaseQty, uint128 qty, uint24 poolTip,
+         uint256 poolIdx, bool isBuy, bool inBaseQty, uint128 qty, uint16 poolTip,
          uint128 limitPrice, uint128 minOutput, uint8 reserveFlags) =
             abi.decode(input, (address, address, uint256, bool, bool,
-                               uint128, uint24, uint128, uint128, uint8));
+                               uint128, uint16, uint128, uint128, uint8));
         
         return swapExecute(base, quote, poolIdx, isBuy, inBaseQty, qty, poolTip,
                            limitPrice, minOutput, reserveFlags);
